Modified to write and donwload labels

+++ Crawler/Kinetics/download.py
@@ -15,9 +15,9 @@ import pandas as pd
 
 def create_video_folders(dataset, output_dir, tmp_dir):
     """Creates a directory for each label name in the dataset."""
-    # print(output_dir)
     if 'label-name' not in dataset.columns:
         this_dir = os.path.join(output_dir, 'test')
+        dataset["label-name"] = [0 for x in range(len(dataset))]
         if not os.path.exists(this_dir):
             os.makedirs(this_dir)
         # I should return a dict but ...
@@ -28,18 +28,11 @@ def create_video_folders(dataset, output_dir, tmp_dir):
         os.makedirs(tmp_dir)
 
     label_to_dir = {}
-    csv_out_file = open(os.path.join(output_dir,"dataset.csv"),"w")
-    csv_writer = csv.DictWriter(csv_out_file,fieldnames=["path","label"])
-    # print(dataset['label-name'].unique())
     for label_name in dataset['label-name'].unique():
         this_dir = os.path.join(output_dir, label_name)
         if not os.path.exists(this_dir):
             os.makedirs(this_dir)
-        csv_writer.writerow({"path":os.path.abspath(this_dir),"label":label_name})
-        # print("writing to csv")
         label_to_dir[label_name] = this_dir
-    # print("csv_writing done")
-    csv_out_file.close()
     return label_to_dir
 
 
@@ -102,17 +95,19 @@ def download_clip(video_identifier, output_filename,
         except subprocess.CalledProcessError as err:
             attempts += 1
             if attempts == num_attempts:
-                return status, err.output
+                return status, err.output.decode("utf-8")
         else:
             break
 
     tmp_filename = glob.glob('%s*' % tmp_filename.split('.')[0])[0]
     # Construct command to trim the videos (ffmpeg required).
+    # Example command
+    # ffmpeg -i /tmp/kinetics/11d602ad-b859-434a-8770-d120dd96f348.mp4 -ss 0 -t 10  -c:v libx264 -c:a copy -threads 1 Dataset/validate/washing feet/--GkrdYZ9Tc_000000_000010.mp4
     command = ['ffmpeg',
                '-i', '"%s"' % tmp_filename,
                '-ss', str(start_time),
                '-t', str(end_time - start_time),
-               '-s', '-1:"%s"' % out_height,
+               '-filter:v scale="trunc(oh*a/2)*2:%s"' % out_height,
                '-c:v', 'libx264', '-c:a', 'copy',
                '-threads', '1',
                '-loglevel', 'panic',
@@ -122,7 +117,10 @@ def download_clip(video_identifier, output_filename,
         output = subprocess.check_output(command, shell=True,
                                          stderr=subprocess.STDOUT)
     except subprocess.CalledProcessError as err:
-        return status, err.output
+        print("Exception!")
+        print(output.decode("utf-8"))
+        print(command)
+        return status, err.output.decode("utf-8")
 
     # Check if the video was successfully saved.
     status = os.path.exists(output_filename)
@@ -134,15 +132,16 @@ def download_clip_wrapper(row, label_to_dir, trim_format, tmp_dir):
     """Wrapper for parallel processing purposes."""
     output_filename = construct_video_filename(row, label_to_dir,
                                                trim_format)
+    
     clip_id = os.path.basename(output_filename).split('.mp4')[0]
     if os.path.exists(output_filename):
-        status = tuple([clip_id, True, 'Exists'])
+        status = tuple([clip_id, True, 'Exists', output_filename, row['label-name']])
         return status
 
     downloaded, log = download_clip(row['video-id'], output_filename,
                                     row['start-time'], row['end-time'],
                                     tmp_dir=tmp_dir)
-    status = tuple([clip_id, downloaded, log])
+    status = tuple([clip_id, downloaded, log, output_filename, row['label-name']])
     return status
 
 
@@ -176,12 +175,10 @@ def parse_kinetics_annotations(input_csv, ignore_is_cc=False):
 
 def main(input_csv, output_dir,
          trim_format='%06d', num_jobs=24, tmp_dir='/tmp/kinetics',
-         drop_duplicates=False):
+         drop_duplicates=False, total=-1, label_csv=None):
 
     # Reading and parsing Kinetics.
-    # print("parsing")
     dataset = parse_kinetics_annotations(input_csv)
-    # print("parsed")
     # if os.path.isfile(drop_duplicates):
     #     print('Attempt to remove duplicates')
     #     old_dataset = parse_kinetics_annotations(drop_duplicates,
@@ -191,27 +188,51 @@ def main(input_csv, output_dir,
     #     print(dataset.shape, old_dataset.shape)
     #     dataset = df
     #     print(dataset.shape)
-
+    if total == -1:
+        total = len(dataset)
+    dataset = dataset[:total]
+    if label_csv is None or not os.path.exists(label_csv):
+        import requests
+        import io
+        print("Downloading csv")
+        url = 'https://gist.githubusercontent.com/willprice/f19da185c9c5f32847134b87c1960769/raw/9dc94028ecced572f302225c49fcdee2f3d748d8/kinetics_700_labels.csv'
+        r = requests.get(url, allow_redirects=True)
+        labels_dict = {y:x for x,y in csv.reader(io.StringIO(r.content.decode("utf-8")))}
+        # print(labels_dict)
+    else:
+        with open(label_csv,"r") as labels_file:
+            labels_dict = {y:x for x,y in csv.reader(labels_file)}
+    labels_dict[0]=0
     # Creates folders where videos will be saved later.
     label_to_dir = create_video_folders(dataset, output_dir, tmp_dir)
-    # print("folders_created!")
     # Download all clips.
+    csv_file = open(os.path.join(output_dir, "dataset.csv"), "w")
     if num_jobs == 1:
         status_lst = []
-        for i, row in dataset.iterrows():
+        for i, row in tqdm(dataset[:total].iterrows(),total=total):
             status_lst.append(download_clip_wrapper(row, label_to_dir,
                                                     trim_format, tmp_dir))
     else:
         status_lst = Parallel(n_jobs=num_jobs)(delayed(download_clip_wrapper)(
             row, label_to_dir,
-            trim_format, tmp_dir) for i, row in tqdm(dataset.iterrows(),total=len(dataset)))
+            trim_format, tmp_dir) for i, row in tqdm(dataset[:total].iterrows(),total=total))
 
+    csv_writer = csv.DictWriter(csv_file, fieldnames=["path","label"])
+    for status in status_lst:
+        if status[1]:
+            # print("writing row...")
+            csv_writer.writerow({"path":os.path.abspath(status[3]), "label":labels_dict[status[4]]})
+    csv_file.close()
     # Clean tmp dir.
-    shutil.rmtree(tmp_dir)
+    if os.path.exists(tmp_dir):
+        shutil.rmtree(tmp_dir)
 
     # Save download report.
     with open('download_report.json', 'w') as fobj:
-        fobj.write(json.dumps(status_lst))
+        try:
+            fobj.write(json.dumps(status_lst))
+        except:
+            print(status_lst)
 
 
 if __name__ == '__main__':
@@ -222,12 +243,17 @@ if __name__ == '__main__':
                          'YouTube Identifier,Start time,End time,Class label'))
     p.add_argument('output_dir', type=str,
                    help='Output directory where videos will be saved.')
+    p.add_argument('--label_csv', type=str, default=None,
+                   help=('CSV file containing the following format: '
+                         'Id, Class label'),)
     p.add_argument('-f', '--trim-format', type=str, default='%06d',
                    help=('This will be the format for the '
                          'filename of trimmed videos: '
                          'videoid_%0xd(start_time)_%0xd(end_time).mp4'))
     p.add_argument('-n', '--num-jobs', type=int, default=24)
     p.add_argument('-t', '--tmp-dir', type=str, default='/tmp/kinetics')
+    p.add_argument('--total',type=int, default=-1)
+
     p.add_argument('--drop-duplicates', type=str, default='non-existent',
                    help='Unavailable at the moment')
                    # help='CSV file of the previous version of Kinetics.')
